+++ src/main/java/com/soklet/internal/microhttp/Options.java
@@ -5,7 +5,7 @@ import java.time.Duration;
 public record Options(String host, int port, boolean reuseAddr, boolean reusePort, Duration resolution,
                       Duration requestHeaderTimeout, Duration requestBodyTimeout, Duration responseWriteIdleTimeout,
                       int readBufferSize, int acceptLength, int maxRequestSize, int maxHeaderCount,
-                      int maxRequestTargetLength, int maxConnections, int concurrency) {
+                      int maxHeadersSize, int maxRequestTargetLength, int maxConnections, int concurrency) {
 
     public static OptionsBuilder builder() {
         return OptionsBuilder.newBuilder();